Remove get_source use for security reasons.
In most cases where checking who is invoking the contract we use get_sender, which is
the address of who called the entry point directoy. The get_source API is who originated
the transaction, and is generally considered unsafe, as someone might make a innocent
seaming contract that then calls your entry point, and thus could trick someone into
doing something, as get_source would be the original person, not the malicious contract.